requirements.txt
absl-py==2.1.0
addict==2.4.0
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
ftfy==6.2.0
future==1.0.0
grpcio==1.62.1
huggingface-hub==0.22.2
idna==3.6
imageio==2.34.0
imgaug==0.4.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
kornia==0.6.12
lazy_loader==0.4
lmdb==1.4.1
lpips==0.1.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
openai-clip==1.0.1
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyiqa==0.1.7
pyparsing==3.0.9
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-image==0.22.0
scipy==1.11.1
seaborn==0.12.2
shapely==2.0.3
six==1.16.0
stack-data==0.6.3
sympy==1.12
tb-nightly==2.17.0a20240409
tensorboard-data-server==0.7.2
tifffile==2024.2.12
timm==0.9.16
tomli==2.0.1
torch==1.12.1+cu113
torchvision==0.13.1+cu113
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
yapf==0.40.2
zipp==3.18.1
